refactor: refactor shared module
