+++ src/app/shared/animations.ts
