+++ src/app/shared/data-table/index.ts
@@ -1,7 +1,6 @@
 export { DataTable } from './models/data-table';
-export { createButton, DataTableButton } from './models/data-table-button';
-export { createColumn, createLocalizedColumn, DataTableColumn } from './models/data-table-column';
-export { createRouteButton, createRouteButtonWithParams, DataTableRouteButton } from './models/data-table-route-button';
-export { createOptions, DataTableOptions } from './models/data-table-options';
-export { DataTableComponent } from './data-table.component';
+export { createButton } from './models/data-table-button';
+export { createColumn, createLocalizedColumn } from './models/data-table-column';
+export { createRouteButton, createRouteButtonWithParams } from './models/data-table-route-button';
+export { createOptions } from './models/data-table-options';
 export { DataTableModule } from './data-table.module';
