+++ src/app/shared/index.ts
@@ -3,6 +3,8 @@ export { Scrollable } from './common/models/scrollable';
 export { switchCase } from './common/util/func.util';
 export { fromSlug, toSlug } from './common/util/seo.util';
 export { LoadingOverlayComponent } from './common/components/loading-overlay/loading-overlay.component';
+export { MenuGroupComponent } from './common/components/menu/menu-group.component';
 export { CommonModule } from './common/common.module';
+export { routeAnimation } from './animations';
 export { ERROR__NO_PAYLOAD, HTTP_CLIENT__MAX_RETRIES } from './constants';
 export { SharedModule } from './shared.module';
